IDE 清理 补充缺失的工具类（部分）
同时，彻底替换掉了 ICOfflinePlayer/ICPlayer 后续适配还需要额外的工作

+++ src/main/java/com/loohp/interactivechatdiscordsrvaddon/main/GUIMain.java
@@ -314,7 +314,7 @@ public class GUIMain {
                 if (downloadResult) {
                     label.setText("<html>Downloading libraries: (4 of 4)<br>Downloaded library \"" + jarName + "\"<html/>");
                 }
-                progressBar.setValue((int) (percentage * 100));
+                progressBar.setValue(percentage * 100);
             });
             label.setText("<html>Done!<html/>");
             future.complete(null);
